requirements.txt
numpy
pillow
openai
wikipedia-api
wikipedia
pydantic
tiktoken
tweepy
requests-oauthlib
